OverlayHandler singleton osztály
A singleton viselkedés garantálja, hogy a főszálban jön létre a példány. Nem kerül háttérbe, mint eddig.
