+++ App/Config/PredefinedActionMap.json
@@ -1,6 +1,7 @@
 {
     "Egérmód ki/be": "self.toggleMouseMode()",
     "Böngésző elindítása": "os.system('start http://www.google.com')",
+    "Feladatkezelő": "os.system('start taskmgr')",
     "Képernyőkép készítése": "pyautogui.press('printscreen')",
     "Számológép indítása": "os.system('start calc')",
     "Gmail megnyitása": "os.system('start http://www.gmail.com')",
