+++ App/main.py
@@ -15,6 +15,7 @@ import Controllers.CameraOptionsController
 from Models.RecognizerHandler import *
 from Models.MediaPipeHandler import ImportHandler
 from Models.Recorder import Recorder
+from Models.OverlayHandler import OverlayHandler
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'Controllers')))
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'Models')))
@@ -45,6 +46,8 @@ class MainWindow(QMainWindow):
         self.stacked_widget.addWidget(self.teach_menu)
         self.stacked_widget.addWidget(self.new_gesture_wizard)
         self.stacked_widget.addWidget(self.camera_options)
+
+        self.overlay = OverlayHandler.getInstance()
 
         self.ml = ImportHandler()
         self.rl = RecognizerHandler.getInstance()
